move some types
