--- packages/uploadthing/src/sdk/types.ts
@@ -1,49 +1,0 @@
-/**
- * A subset of the standard RequestInit properties needed by UploadThing internally.
- * @see RequestInit from lib.dom.d.ts
- */
-export interface RequestInitEsque {
-  /**
-   * Sets the request's body.
-   */
-  body?: FormData | ReadableStream | string | null;
-
-  /**
-   * Sets the request's associated headers.
-   */
-  headers?: [string, string][] | Record<string, string>;
-
-  /**
-   * The request's HTTP-style method.
-   */
-  method?: string;
-}
-
-/**
- * A subset of the standard Response properties needed by UploadThing internally.
- * @see Response from lib.dom.d.ts
- */
-export interface ResponseEsque {
-  ok: boolean;
-  /**
-   * @remarks
-   * The built-in Response::json() method returns Promise<any>, but
-   * that's not as type-safe as unknown. We use unknown because we're
-   * more type-safe. You do want more type safety, right? 😉
-   */
-  json<T = unknown>(): Promise<T>;
-  text(): Promise<string>;
-
-  clone(): ResponseEsque;
-}
-
-export type MaybeUrl = string | URL;
-
-/**
- * A subset of the standard fetch function type needed by UploadThing internally.
- * @see fetch from lib.dom.d.ts
- */
-export type FetchEsque = (
-  input: RequestInfo | MaybeUrl,
-  init?: RequestInit | RequestInitEsque,
-) => Promise<ResponseEsque>;
